# cienergytool/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: cienergy Ingester API
  version: 1.0.0
  description: |
    HTTP API for the cienergy ingester. Accepts SCI-compliant energy reports
    (see [`docs/schema/v1.json`](../schema/v1.json)) and stores them in Postgres
    for the Grafana dashboard.

    Authentication: if `INGESTER_TOKEN` is configured server-side, all `/v1/*`
    routes require `Authorization: Bearer <token>`. Health endpoints are
    always public.
  contact: { name: cienergy maintainers, url: https://github.com/axa-oss/cienergytool }
  license:  { name: Apache-2.0, url: https://www.apache.org/licenses/LICENSE-2.0 }

servers:
  - url: http://localhost:8080
  - url: https://cienergy.example.com

security:
  - bearerAuth: []
  - {}        # also allow no auth when INGESTER_TOKEN is unset server-side

paths:
  /healthz:
    get:
      summary: Liveness probe
      security: []
      responses:
        '200': { description: alive }
  /readyz:
    get:
      summary: Readiness probe (DB ping)
      security: []
      responses:
        '200': { description: ready }
        '503': { description: not ready }

  /v1/runs:
    post:
      summary: Ingest one energy report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'https://axa-oss.github.io/cienergy/schema/v1.json'
            examples:
              minimal:
                $ref: '#/components/examples/MinimalReport'
      responses:
        '201':
          description: Stored. Returns the run id and computed totals.
          headers:
            Location:
              description: URL of the created resource
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:        { type: string }
                  sciValue:  { type: number }
                  totalCO2:  { type: number }
                  energyKWh: { type: number }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
    get:
      summary: List recent runs
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        '200':
          description: A page of runs ordered by `startedAt` DESC.
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:           { type: string }
                        repository:   { type: string }
                        workflow:     { type: string }
                        startedAt:    { type: string, format: date-time }
                        energyKWh:    { type: number }
                        totalGCO2eq:  { type: number }
                        sci:          { type: number }
                        zone:         { type: string }
                        arch:         { type: string }

  /v1/runs/{id}:
    get:
      summary: Fetch one full report by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The full SCI report JSON, as ingested.
          content:
            application/json:
              schema:
                $ref: 'https://axa-oss.github.io/cienergy/schema/v1.json'
        '404': { description: Not found }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque

  responses:
    BadRequest:    { description: Malformed JSON or query params,
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized:  { description: Missing or invalid Bearer token,
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    TooLarge:      { description: Request body exceeds MAX_BODY_BYTES,
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unprocessable: { description: Schema validation failed,
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    ServerError:   { description: Unexpected server error,
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:  { type: string }
        detail: { type: string }

  examples:
    MinimalReport:
      summary: A minimal valid v1 report
      value:
        $schema: https://axa-oss.github.io/cienergy/schema/v1.json
        specVersion: '1.0.0'
        sciSpecVersion: 'ISO/IEC 21031:2024'
        run:
          id: gha-7001
          platform: github-actions
          repository: axa/claims-api
          workflow: ci.yml
          commitSha: a1b2c3d4e5f60718293a4b5c6d7e8f9012345678
          startedAt: '2026-06-01T08:12:33Z'
          endedAt:   '2026-06-01T08:19:07Z'
          durationSeconds: 394
        runner:
          os: linux
          arch: x86_64
          vcpu: 4
          ramGiB: 16
          cpuModel: Intel Xeon Platinum 8370C
          tdpWatts: 270
          provider: github-hosted
          region: US-VA
        energy:
          totalKWh: 0.020966
          byStep:
            - { name: build, kWh: 0.012285, durationSeconds: 210, source: eco-ci-model, cpuUtilPct: 78 }
        carbon:
          operationalGCO2eq: 8.198
          embodiedGCO2eq: 0.920
          totalGCO2eq: 9.118
          gridIntensity:
            valueGCO2eqPerKWh: 391
            source: electricitymaps
            zone: US-VA
            timestamp: '2026-06-01T08:15:00Z'
          embodiedSource: boavizta
        sci:
          value: 9.118
          unit: gCO2eq
          functionalUnit: '1 pipeline run'
          R: 1